+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Redux URLs.
 - Update expenses in the budget.
 - List expense categories associated in the budget.
+- Cleanup the DB when a upload does not have expenses associated.
 
 ## 2024-01-21
 - Expenses CRUD forms.